daemon: extract legacy-config migration into its own module
Continues #41. Pulls the legacy-shape migration code out of `mod.rs`
into `config/migration.rs`:

- `maybe_migrate_legacy_connections` — `[llm]` → `[connections.default]`
- `maybe_migrate_legacy_purposes` — `[llm]`/`[backend_tasks.llm]` →
  `[purposes.*]`
- `pick_free_connection_id`, `pick_backup_path`,
  `file_has_top_level_table` — supporting helpers

`load_daemon_config` calls into the module via `migration::…`. The
parent test module reaches `file_has_top_level_table` and
`pick_backup_path` (both `pub(super)`) for the migration-detection
and backup-path-escalation tests; everything else stays private.

Drops the orphaned `connection_from_legacy_llm` import and the
`ConnectionRef` / `ModelRef` / `PurposeConfig` imports from
`mod.rs` — those moved with the migration code.

`config/mod.rs`: 3536 → 3242 lines. New `migration.rs` lands at 346
lines. Cumulative for #41: `mod.rs` 4373 → 3242 (-1131 lines), now
plus five focused submodules.
